Provides a client C API and a CLI tool
Adds a client C API to allow client programs to connect to a Memo server
and publish and subscribe messages. A message loop starts and listens
for events and dispatches messages to callback handlers when received.
Clients can specify their own allocator function to allocate memory for
messages if needed.

Included is the memo command line tool that uses the new API.
Sub-commands allow users to publish and subscribe.