Fix memory havoc
Previously Memo just leaked memory everywhere. This PR addresses most of
this leakage. A series of 4K buffers are registered with io_uring. These
pool buffers are used to read messages and are considered checked out of
the pool until all writes from it are complete. If the buffer isn't big
enough for a message then a msg_buf size buffer is specially allocated
and freed when all writes are completed.